factor out `ApiDecodedTransaction` from `Api.Types`
